Cast RNN-T logits to float32 before torchaudio loss.
torchaudio rnnt_loss rejects bfloat16 under bf16-mixed training; upcast logits at the loss boundary only.

+++ src/SpeechToText/models/common/rnnt.py
@@ -190,7 +190,7 @@ def rnnt_loss_mean(
     ).to(dtype=torch.int32)
 
     loss_any = rnnt_loss(
-        logits=logits,
+        logits=logits.float(),
         targets=targets_2d,
         logit_lengths=out_lengths.to(dtype=torch.int32),
         target_lengths=target_lengths.to(dtype=torch.int32),
